chore: refactor pubsub into pipe, move request & response to helper.ts
